Add files via upload
Analyzing 3 years of stock data with VBA code. ~750,000 rows per year. Took 4.5hrs to run on a 2020 macbook pro with M1 chip. This is one code block to execute all the requirements, except for converting the date to numbers (the original excel document had the date numbers stored as text).